+++ class-app/internal/api-student/dto/requests/student.go
@@ -1,17 +1,17 @@
 package requests
 
 type StudentCreate struct {
-	ClassId int    `json:"class_id"`
+	ClassId int64  `json:"class_id"`
 	Name    string `json:"name"`
 	Email   string `json:"email"`
 }
 
 type StudentUpdate struct {
-	ClassId int    `json:"class_id"`
+	ClassId int64  `json:"class_id"`
 	Name    string `json:"name"`
 	Email   string `json:"email"`
 }
 
 type StudentId struct {
-	ID int `json:"id"`
+	ID int64 `json:"id"`
 }
